requirements.txt
# Core dependencies (compatible versions)
crewai>=0.80.0
langchain-google-genai>=0.0.5
PyGithub>=2.1.1
python-dotenv>=1.0.0
google-generativeai>=0.3.2
pyyaml>=6.0.1

# Fuzzy matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# Optional - only if you need Solidity compilation
# py-solc-x>=2.0.2
